build fixes, improvements to enablePartialValues mode in parser
Robustified the build and install system so you can run sc from the bin or main directories
fixing bugs.

Parser enablePartialValues mode works in many more cases.  We can support code-completion and
parsing of the UnitConverter swing sample in rtext.  Big step towards parselets support for general
IDE-like development.  The main missing parts are the UI integration for various IDEs and the
ability to expand a parser error so that it skips either elements in the model or text in the file
to resume parsing.