more partial values fixes
workaround for awful chrome hang bug - it seems to happen when we use the string object's class in a Java hashtable.
Possibly we are storing the sc$id in the object and that hangs?  Or maybe it does not stick?  Or maybe we are using the
hashCode and equals method on the prototype, not the class and it goes into lala land.

QuestionMarkExpression's get completed in code.  this involved a fix for choosing the best error when in partial values
mode.