performance improvements to the parser, particularly for HTML Language
added a cacheResults property on parselets which you can set on those
parselets that might be parsed more than once at the same location in the file.

This speeds up the HTMLLanguage which parses the tags first as a tree and
secondly as a simple tag.  It's not the ideal fix but nice that the changes
to the grammar are minor.

Used Symbol(repeat) instead of Sequence((''), REPEAT) to reduce memory usage
and speed up varIdentifiers and identifiers in general.  This involved fixing
the Symbol(repeat, ANYCHAR) case which we have not used before.

Minor tweaks to various accept methods for minor speedups.

when ENABLE_STATS is on we also generate a table which counts the # of parse
operations for each parselet for each position in the file.  If you see a parselet
being used on the same location more than once, it's worth adding cacheResults=true
most likely so the second one is just a hash-table lookup.  It's true that in general
this occurs when you replicate the grammar in a "choice" node.

Use indexed choice for the template language and html tag