three bugs but still not working
now that data binding expressions are inside of element tags apparently we need to do the findMethod in the element so that we can resolve types in the attribute itself for good syntax highlighting.

fixed weird case that showed up in JS with parameterized types in the dependency list

fixed even weirder wrong test for the Modified class started method.