inactiveLayers are now ordered
otherwise, we'll keep swapping types when two types extend the same base
type but not each other.  We need to create a consistent ordered stack
so that all of the layers we need will be present.  We'll need to expose
this inactive layers list so users can remove entries from it as there's a
chance of conflicts between unrelated layers you are editing.