inactive layers fixes and packing improvements for android
Brought over some zip code from apache's archive utils for extracting the
extra data in a zip file in a failed attempt to extra permissions properly.
Instead no longer using any of the internal unjar stuff and instead just
use unzip from exec.  We could find the extra data in each file ent but it only
had the user ids and last modified times - no permissions.  I think that data is
in the header extra data but I could not find that in the Java api.  So all of this
code can be removed but we are archiving it here just in case it's useful in the
future.

fixes so that we can load up multiple runtimes for android, gwt, etc. to at least parse
all of the code at the same time.

Layers can be excluded - i.e. they are still in the baseLayers but not active in this runtime.

Layers can be disabled - i.e. they are just not started... this is how the old inactive layers worked but
now that we are starting the inactive layers, all of the doc references wanted to be resolved which let to lots
of errors and the desire to have all of the runtimes running at the same time, even if they are inactive.

new fork method for android to run the emulator