ide feature complete - regression tests about 50% tested
added @TypeSettings(bindableProps) so that we can mark properties as bindable in the generated code even if we do not code-gen a property for them.  An alternative to this fix is to just load those classes in src description but it seems like avoiding these runtime warnings will be worth the effort to code-gen this stuff out like we are doing now.

we did not yet support this for annotation layers e.g. javax.swing.JComponent's @Constant propertiesa.   The useRuntimeReflection property should be made per-type and we should dust it off and make sure it still works (it was added for GWT which has been deactivated for now).

allInheritedAnnotations so that we can implement

track whether or not a model has been added to the type system and do that lazily.  The IDE might load the model for us in a new code path so we add it the first time the type system requires it.

bug fixes for debugging registration

fix for bound properties during rebuild - we were not collecting them except on the transformed model for some reason?

added 'transient' flag to speed up semantic node traversal and ensure we only
traversed the seamntiuc nodes to avoid potential oddities. Caught a few bugs where deepCopy relied on the missing transient flag.

when retrieving an inner type, we may find a .class file when we need a source file because we initially resolved the base class as a ocmpiled class.  We handle this now by updating the extends type at that time.  I suppose also could have just fetched the src type when we initialized the type but that seemed like it might undo an optimization.

scope and definition processors are registered on the layers now so that we can look them up using dependencies.

another case in ModelUtil is now using GenericType instead of getTypeDeclaration so that we preserve the param type info.  This uncovered a few other bugs where we needed to check for Parameterized types in the type system.

treating block statement as a "top level statement" for debugging purposes and the convention will be that we use it to represent the last statement of the method.

reset all build info when clearing all active layers