added "replacesLayer" for improved control over layer ordering
html.schtml now marks html.core as a 'replacesLayer' which means anyone who depends on html.core now depends on html.schtml when it's included.  This moves html.schtml ahead of js.core when they are both together.  It also ensures that when js.core is built, html.schtml is included due to this dependency even though js.core can be used without the tags introduced in html.schtml.

It turns out this is going to be an important feature for those side-cases where the type hierarchy does not imply a strict layering ordering.  You need to essentially replace that layer in the ordering which means be in front of any layer which depends on the previous layer.