Added typeIndex and process filters for layers
The type index is maintained always but only initialized when running
in the IDE right now.  It stores the set of types defined in each layer the
and the base types for quick high-level lookups.

It's currently missing inner types, fields, properties, and methods.

Just as you can separate layered systems with runtime meta-data, you can now also do it with process meta data.  It works in a backwards compatible way - so layers can define runtime constraints and/or process constraints.  At least one process is created for each runtime and each runtime and each process has a runtime processor.  Each runtime has a default 'null' process but when you add a process for that runtime, we'll use the default layered system for the first one.

When you define either a process or runtime constraint on a layer it controls which layered systems that layer can be put into.