disabled used of ProcessIdent as RuntimePrefix
we need consistent mapping of layer to build-dir... adding layers can remap the default process name and change
the build dir from say java to java_Desktop when the swing layers are added.  maybe we should have a flag for
multi-process VM, use the default only for one process (and mark it so it's consistent).