bug fixes for multi-runtime systems
support the case where none of the active layers exist on the core layered system.

for ensuring the start layers, we always iterate over the peerSystems - not just in peer mode, since
buildSystem may not start your dependent layers if it's a peer-to-peer system

initSync bugs with this case.  By default, do not sync from the same runtime to itself unless you have configured to sync the two processes.
for consistency we should probably explicitly enable web-java to browser-js

fixed bug with getAnnotatedModel - so it can find the models if started from a peer system and vice versa

new apis for getting the active model - for debugging so you can pick the model for a given runtime.  Might need to move this to the process level...

checkLayerPosition - diagnostic code for ensuring the integrity of the layered system's position

properly preserve current layered system around initruntimes

reorg layered system and layer fields so the main ones are easier to find in the debugger and source code - roughly in "priority" order