added external class path for android java.lang.AutoCloseable conflict
Unfortunately a class loading conflict came up where we cannot load android.jar into a ClassLoader.  The Notes test pulled in this class and there's a security exception we can't disable.

Android now puts android.jar into the "externalClassPath" and that's use to load CFClasses to compile the type system.

Still support the old "cross compile" option but that seems less good overall because we can't separate which classes are loaded which way.

Fixed bug with server rendering of UC demo when there's no select body.

incremental compile bug - we can't consider a JavaModel which was reloaded to be annotated as a changed model or else it messes up the type group computation (because we never process that model).

Fixed several bugs with the CFClass stuff... there are still wholes in there - the biggest of which is more accurate parameterized types.