several bugs with IDE related features, include/exclude layers
A couple of subtle problems with includeForProcess - the predence of the constraints matters
when you have one layer extending a base layer with runtime and process constraints.  Need to pick up
the right constraint.

Fixes for build separate layers

Code to recalculate layer position for build separate layers was off

buildSeparate layers not part of the specified layer but do need to have
that include test done each time anyway.

For some reason we were cleaning the buildSrcIndex - i.e. removing all build files, every time we saved
the build info.  That clean operation is only done now for the initial complete build.