added disabled runtimes and fixed bugs found during ide type index creation
Each layered system can mark certain runtime names as "disabled".   We automatically then
disable the layers in that runtime and no LayeredSystems are created.  This makes it easier to
manage the current layers directory, for example.  Installing android is not easy and should not
be required to get started with SC.

Refactoring to the package manager so that we can install layers by default in the IDE - before
there's a layered system.

Bug fix for a 'class x' which overrides another type 'x' in a base layer.  It now sets replacedByType in the
base type so that subsequent layers can resolve the right type, even if they pre-resolved it earlier.  This
does not happen for active layers because we always create all layers up front and resolve from the top of the
stack, working on down, but for inactive layers, it's incremental.   If we define a class which uses UIColor, then we
replace UIColor in a subsequent layer, when we use one of those old references in a layer downstream of the new definition
it needs to be able to resolve the new definition.

Element for inactive types - fix to a recent earlier fix.  Need to eliminate the explicit extends type when it's a modify type
or we generate a modifyInherited type which extends another type - that prints an error later on in the ModifyDeclaration.