fix for using replacedByType for replacing one class with another only enabled for inactive layers
In the active case, we resolve types from top-down so we always get the most specific version of a type we
need.

For the inactive case, we need a way to track that we've replaced a type at a particular layer, then change the
'resolve type' to use the ref-layer so we return the appropriate type for that layer.  Right now, the inactive case
will just choose the most specific type for all subsequent resolved references.  for the index case, this works because
we work bottom up in the reference graph.