@Sync changes and repeatWrapper enhancement
defaultSyncMode is now deprecated.  Instead use the @Sync annotation on the layer itself.

includeSuper is now supported and it defaults to false.  Now you have to explicitly add includeSuper
if you want to synchronize the sub-types properties.

The repeat wrapper takes the old tag object as a parameter.  It can just return that if you are able to reuse the
same tag object for the new value.  This allows the repeat wrapper hook to be used for customizing how the
element is created for a particular slot.