locking and other IDE fixes
Mark the model as temporary when IntelliJ parses the file with the
IntelliJIdeaRulezzz token for completion.  This avoids us from
trying to resolve those models (i.e. from getReferences in the PsiElement)

Diagnostic for type index  and sync

Fixed a couple of type parameter bugs - when we can't resolve a type
don't add the ClassType - null means wildcard as a type anyway

Methods were not resolving return type correctly in some situations