fixed bug IDE in resolving type parameters
The code to generate a ClassType from a Class or java.lang.reflect.Type
was not using the "refLayer" argument - i.e. the context to find a dependent
type in the IDE.  It was not finding types defined in the source layers.

Also fixed exit status of the 'sc' script - hopefully now the test scripts
will exit on the first failure.

Turned off the 'retry' after failed compile.  It does not work yet and causes
some frustration when it pops up unexpectedly.