Improvements to type system, inferred types and in general
Note: full regressions have not been run.  This checkin is just
getting too big.

Need a new flag for finding static methods to deal with the
case where there are static and non-static methods with the
same signature.

Added 'bound' flag for retrieving parameter types either
in the format in which they are defined or in the context
of the method invocation.

Generilized HTML language for XML

More supported for lambda's

Support for differentiating ? super X