Fixed many bugs - still not through regressions
HTML language - static variables not assigned before instance vars?   Apparently that's possible!

Constructor outer instancae parameter needs to be taken into account for propagating inferred types when
binding to compiled methods.

Choosing methods with varArgs - behavior had to change.  There was a case where the method with more parameters
was the right match.  Now the number of args is only used to choose the default if there's an ambiguous match.
We should print an error in this case like Java but living with Java handling that one for now.

addRunCommand - can't support the signature (String, String...) as it's ambiguous and behavior changed recently in
our "pickMoreSpecificMethod" so this one broke.  Just removed that api.

JavaType.getTypeDeclaration - behavior changed with type parameters.  We now by default return the type variable but
the JS code - getAbsolutePathName - in particular required the old behavior.  Now there's an arg to that method to resolve
to the default value or not.

TypeMethodExpression - completely reworked so that it always adds it's typedIdentifier into "getAllIdentifiers".  We use that
now - for simple identifier expression, this just returns identifiers for now we can treat both types with the same code base.
Currently the actual method type parameters are ignored but this should be easy - just create a new ParamTypedMethod for that
situation when they are there.  So far, we have no tests which need those type parameters for validation so we'll wait till we
hit a case.

Fixed injection for HTMLLanguage