support for layer fields and inner-objects
Each layer now implements IDynObject and can now define fields and
sub-objects.

The extends layers do not contribute to the set of fields in the layer
as in a regualr class.  Instead, we search the layer type hierarchy and
find those fields in the sub-layers.  This lets one layer easily modify
the fields set in downstream layers with a simple property assignment.

This required a few changes to how the property cache is built, how
the dyn object implementation works for layers, and now for synchronization
the layer type needs to get out of the way - so we still just send across
the original fields in sc.layer.Layer.

Fixed a sync bug where 'createdLayers' was built on just the base name when
it needed to be the full type name.