fixed bugs with 'dynamicNew' option for dyn types
Turn on liveDynamicTypes for layers, even if they disable it
since we need the inner/outer type map for the dynamicNew case.

Added a useDynamicNew option and improved the support for that
optimization.  It lets us avoid the dynamic stub for simple cases
like creating a new object and setting properties.   As soon as
someone adds a field or method in the type hierarchy though, we
generate the stub for the first dynamic class in the chain.

LayerComponent - both Language and LayerFileProcessors extend
this layer component class which is used to hold the parent layer
(if any).   We set propagateConstructor on this class so we can
construct it as an object inside of a layer and have that value
provided automatically.  Unless we want the main SC build to modify
the source code and add that constructor, we needed the Layer
constructor in each language.