Warning - checkpoint commit - regressions not run so may not be stable!
maven fixes - put files in directory based on version number.  otherwise, we
may find an older cached .pom file with the wrong version and use it without
refreshing.  This is particularly a problem when using the same repository for
projects with conflicting requirements.  we should not be using different versions
of the same package for the same build.

added buildPrefix to the src path types - so you override the location of
files e.g. change from web to webapps/ROOT for a given src path type.