update type reorg and sync layer bug with same name types
This fixes the -dyn editor.js.mixin test.js.helloWorld test

For update types the reorg fixed one major case of the stopped type getting reinitialized
by ensuring we update all of the data structures for the inner types as well as the outer type
before we start the outer type.  That way all of it's references should point to the new
inner type.

This was a problem all along but now that we are refreshing modify types after compile, it
would pick up the stale inner type that got re-added to the cache afterwards.

It's easy for us to reinitialize the old type after it's been stopped - just by using it from
some stale reference.  Fixed a couple of other sources of stale references:

  - getDeclarationType() called on the old type during updateType (after the stop call).  Either use the new type or now
    we just cache the old declarationType.

  - getDependencies - it's only easy in a couple of places to invalidate the dependencies.  To fix this problem, I had to resolve
    to get the latest references.