fixed a few more bugs from the tests
the resetClassLoader and refreshBoundTypes for classes still has
some problems but this fixes the tests at least.  I think we need to refresh references as we resolve them so we are guaranteed to get the most accurate type.  For now, isAssignableFrom at least matches Classes from different class loaders.

Also need to flush the typeIndex in the JavaModel for the reset of the class loader.

Fixed bug in resolving the root for static types when the class has been optimized away.  Not sure why the field's type is not always the type we want to use here... so that's the way we back this test up when we can't find the enclosing type that matches.