Fixed CFClass bugs with type parameters
When matching parameters for the inferredType method we need to match an unbound type parameter

Fixes to properly resolve the default for a type parameter at the right time.  If it's resolved too soon, it can keep the real value from being set.

various CFClass fixes - 'default' modifier, comparing Comparable and int