Fixed more CFClass bugs
the TypeSignature used for StrataCode needs to have type parameters bound to their defaults

Switch to runtime method and compiled class for a few places where CFClass and CFMethod show up

Ignore inner classes we can't find in the file system.  They might exist in a base-type and for some reason be
listed as inner type names in this class.