java language and maven fixes
- inferredType passed to defineMethod to fix of(T...) and of(T) discrepancy.  Need to
know the types of the type parameters to select the right method and so need to have
inferredType before making the final selection.

- JavaType for methods should be an [] of what's declared for repeating parameters.
It's this way in the compiled class format so we are now doing that for source types as well.

- pickMoreSpecificMethod - subtle semantics for matching array parameters with repeating arguments.
also selecting whether to use (T...) versus (T).  Primitive arrays behave differently than non-primitive
arrays as well.

- Allow +=, -= etc. for int += float

- Better rules for other selfOperator validation

- refreshBoundType with "CLASS" mode breaks when we need to resolve an identifier for the super.x(..)
resolution.  We should probably write a special refresh for just the super case - to refresh the case
type and then get the super type of that.  For now, we have a way to refresh the classes in isAssignableFrom
or compare using class names so that this refresh stuff works, even when dealing with a mix of classes.

- MethodReference type parameter change - now we use the reference method's return type and parameters to
pre-populate the types of the parameters to the new method and return type of the method.  After we build up
the lambda body, we may do that again in a more general with for all lambdas.

- A few misc type parameter fixes for cases that showed up testing more Java code

- Class<?> getClass() method return returns Class<EnclosingType> getClass() through a special rule not defined
in the language spec

- Maven fixes:
   - support parent source modules
   - robustness fixes - save all .ser files after all deps have been resolved so we have consistent object trees
when we restore.
   - put the .ser file into the version root to allow for multiple independent versions at the same time.  This
happens in maven only for the POM files - we may get the wrong version of some module POM included from the parent
that only gets used to resolve variables in the parent POM.

- repository sub-packages

- git repository uses --depth 1 for faster checkouts.  Should make that configurable?