fixed bug with -dyn mode
We made one dynamic stub extend another one in the
type hierarchy.  That should never be necessary because
the first dynamic stub has already exposed the compiled contract
for the type.  Because of how we use a single _super_name
method in the type hierarchy, if we have one stub extend another
it causes an infinite loop trying to resolve... we find the
sub-classes _super_x when we need to get to the compiled super method
and start looking over again from there.