Renaming type, clone models for parseSrcBuffer, misc fixes
First pass at handling the rename of a type without refreshing all indexes.

In the IDE, when we are parsing a buffer model, we also clone the new model
into other runtimes which have that file.  That way, we keep all models up-to-date
with the currently edited file.

Started work at adding the refresh of inactive types but that's not being used now.