Fixes for the IDE, styles with generated types
Collect all modified types of each extends and super type.

Consistently use updateModel/updateType - even for changes to inactive
models so that we have a way to mark the replacedType

When fields change in a base type, increment the member cache
version for all sub-types - even in the inactive case

Parsing with a style adapter that maintains offsets - we must
run through all generated space characters through the style
api or the counts become off.

Bugs with formatting a parent parse node when the format context
does not include the entire document.  Need to walk up the semantic-value
hierarchy and gather enough parse-node info so that ctx.nextChar()
returns the right next character.  Now we go up 2 levels in the
hierarchy and all of the cases we care about seem to work.  The other
way we could fix this is to always generate a format context for the
entire document - then format everything all at once.  Also, when we
refactor nodes, maybe we can use keep the existing spacing so it
is not treated like a generated node.