semantic length, typeIndex refresh, constructor exprs
- to improve error highlighting, the semantic length lets you eliminate the trailing space
for any semantic node.

- continue to improve saving the type index - use last modified to make sure
the type index's last modified time reflects the latest time of the type we've just
modified.

- restore type index - have a way to identify the last "main" system and restore the process ident
for that system quickly - before we start refreshing the layers.   That way, we don't start initializing
layers until we know the process name for the default system.

- identifier expressions for the super(xx) in the constructor now resolve to the
parent's constructor (unless it's the default constructor).  Added errors for the
constructor validation.