restore process/runtime definitions for type index
ProcessDefinition and DefaultRuntimeProcessor are now
serializable so we save the processes and runtimes
that were referenced in the type index when we initialize
it when we read from it.

Otherwise, when restoring the type index we might
initialize a 'java' runtime (which should be java_Server)
and not know to find the java_Server's type index entries.

sync bug with save type index