Various bug fixes - maven support and language fixes
- for maven, better hierarchical POM files - classifier, war, trim fields,
parent-path and module path versus artifactId - really two separate fields
so take that into account

- improved assignable test for arrays when dimensions do not match
turns out Object is assignable to int[] etc. so we need to handle
that and approve those tests

- missed call to isVarArgs which meant that array args matched multiple
args in some cases.

- for iterable<T> in for statements, we need to resolve the value of
the type parameter in a subtlely different way so that required a new
primitive type.