Initial reparse implementation
Essentially reparse is a copy of parse with extra logic
to merge in values from the oldParseNode which have not changed.
It's built right now for only a single 'diff' region to support
the most rapid editing operations.  We could generalize this without too
much trouble down the road but it seems like the second or so delay to
do a global replace in a large file is not a big problem.

Implemented with two passes:

- findStartDiff - compare supplied parse-node with the new text.  Find the
first primitive parse-node which is different in the old and new text.
- findEndDiff - walking from the end of the two files - find the last different
node in the text.  We make sure this node is at least as high in the tree as the first node.  It's common for startDiff and endDiff to be the same node.

- Now we do the parse() logic again but we advance the parse-node down the tree as long as we are in an unchanged region.  As soon as there's a change, we null out the old parse node and reparse.  We stitch in the value and execute the semantic value mapping as we back out.  This has the effect of somewhat selectively updating the old model with any changes made.

- As we exit the parse method, we continually check if we've re-entered the "sameAgain" section - i.e. the part of the end of the file that is identical.  At this point, we start using the oldParseNode again.  We just need to be careful to update the startIndex in the resulting parse-nodes because they may have moved in their offset.

* fixed bug with toString on uninitialized types - use OperationString for a type before it's been initialized.

* semantic values only update if the value has not changed.

NOTE: Down the road we might eliminate the duplicate code between parse and reparse and just use the new reparse logic with an oldParseNode = null.   There are several reasons for duplicating the logic - performance? debugging, and the potential to change the results.

It's nice that this will be only used for the IDE with enablePartialValues mode on so there's a bit more tolerance for an invalid parse.  We will continually reparse the entire document but I see the potential for occasional ambiguities that might require a reparse.   For now, we are prioritizing fast over exact to some degree as we can always re-parse in the normal mode periodically to fix any errors that creep in due to "fixing" an old model that's already been patched over by error skipping.