Type parameter and lambda expression fixes - one maven fix
- addition of "wildcard type declaration" to represent ? as a special type
- addition of method-type-params to the defineMethod as an option.  Used by
identifier expressions which specify method type parameters explicitly - e.g. foo.<a, b>(bar)
we need to use a, b - the method type params in the ParamTypedMethod in this case.
- lambda expressions also populate type parameters when explicit types are specified for the
lambda parameters.
- careful when converting type-declarations to JavaType's.  There are cases where we do re-resolve
the type and parameters because we get more specific values than we are given.  There are other times
we are unable to resolve the type and so must use the type-declaration to initialize things.

- fixed grammar bug with try(resources) {  }

Maven:
  - pre-init sub-packages so they get registered before we look up dependencies - otherwise we have a sub-package
which does not know it's a sub-package and thus can't be re-installed