more param type fixes for running not in cc mode
Now the most difficult type param tests work using native reflection,
not the cross-compile mode.

* ? super T rejected isAssignableFrom for ordinary types

* when using t -> t in lambda's, in some cases the type of 't' must
be derived from the return type of the method.  For now we insert
a cast like - "return (MethReturnType) t;" but perhaps the definition
of the type for the Parameter 't' should be updated with the MethReturnType?

A few more misc minor changes