maven changes and method definition override annotations
- maven now stores package index files in the build layer (as long as one is defined).
Not sure this is entirely right because some files like junit end up using the shared
storage.  But since the package index stores dependencies, and since those might
change for a given version of a given package based on who is building it we can't share
these files between projects.  Ideally we'd be checking the dependency context and
if it's different, we'll need to re-generate the package index.