more reparse fixes
- added "alwaysExtendError" mode - when set, for parselets with
skipOnBodyParselet set for reparse removed the
extra pass to separately parseExtendedErrors.  Now for those parselets we'll
always extend errors when there is a valid exitParselet.

It was hard/expensive to essentially reparse twice and the extra reparse made
it harder to keep the old parse-nodes in tact to do incremental stuff in
the second parse.

It might make sense to do this for the normal enablePartialValues mode but for now leaving that alone since it's not going to be in the performance bottleneck
once the incremental parse stuff is reliable.

- fixed a few bugs with reparsing the chainedTypes property of ClassTypes -
i.e. a repeat sequences with multiple components - e.g. ".", name, optTypeParams
for this pattern the semantic value is only incremented on the last parselet
in the sequence.

- Fixed a couple off-by-one errors in clearParseNodes - need to clear zero
length errors that are on the border and need to clear any parse-nodes
that are before the line (but don't include the line).

- Need to adjust the currentIndex before calling addForReparse (not after)

- For sequences that are reparsed in an incomplete way, we need to cull
any old nodes were not reparsed