Reparse bugs, testing improvements
- Keep track of nodes which are part of an incomplete parse
so that we always reparse these nodes during the reparse.
The errorNode flag gets set on a parent parse-node when it's part
of a partial value, or when it's last component is an error.
We don't advance the 'lastVisitedNode' past these error nodes and
then start reparsing at the last visited node.

Added toString's for more types as we use them in the model errors
and are great for debugging.

More robust "diffNode" method to replace using deepEquals.  Since
reparsed model differences that are inconsequential are relatively
common, we need to ignore specific errors in the testing process.
This happens mostly because we don't remove empty arrays and the reparse
can break down error sequences differently (for some reason I'm not sure).

Fixed a bug with keyword acceptance of non StringTokens.  The string
values cannot be compared against IStrings apparently.

More work updating array values for reparse changes - a null may mean
we need to clear out the array if it's one of those "propagate" value
nodes.