Reparse fixes
- interactive features to TestUtil for accepting new versions

- added alwaysReparse flag to Parselet.  Right now, this is just a signal
to not descent the parse node tree in the getBeforeFirstNode method.  The
beforeFirst node is the node which starts the reparse.  We choose it as
the first non-error node before the first-diff-node... but since that node
can be large, we use the parselet's beforeFirst method to descend larger
elements to choose the last parse-node component that makes sense.

Right now, we only set this on spacing - essentially do not choose the
spacing as the start of the reparse.  Choose the node that includes the
spacing - e.g. the symbol or identifier or whatever in front of it.

- again, improved "anyChanges" to look at the start index of the beforeFirst
node in more cases.  This code could be cleaned up probably as there are some
tests that now seem redundant but it's working and not that complicated or slow.

- fixed bugs with lookahead and reparse nodes

- fixed bug with chained result sequences in progating errors.  That should
be moved to the regular parse method I think but it only involves the
partial value produced in the error which I think we have worked around
the other places it is used.