reparse fixes and other incomplete model checks
- for chainedResultSequences that were being reparsed, need to validate the "skip-child"
value hasn't changed

- for ordered choice sequences, we might have a better error than a valid reparse - e.g.
We reparse a statement "String" (an identifier expression) instead of "String foo = " (a local variable declaration).
We should probably make this change when "enablePartialvalues" is set in the normal reparse.  This might clean up
some of the reparse model differences?

- in some cases we are reparsing a model which has generated nodes - SpacingParseNode instead of the
explicit # of spaces... It would be nice to have a method which just converts the formatting parse-nodes
into regular ones but for now, we're essentially just formatting them as part of the findStart and findEnd
diff methods.  Need to fully test this...