checkpoint - tests not run
- some reparse changes in progress.  currently failing on re23 GenarateUCs40.sc
because the 'null' in place of catch is being accepted cause it's in the
unchanged region.  When we reparse an old parse-node that's smaller we need
to somehow extend the changed region.  Or otherwise determine when a 'null'
can be accepted - by tracking the offset in the old parse-node or re-validating
the null parse.