Turn off updateParseNodes mode for list changes
It doesn't matter for the initial parser (when the semantic value
has not been created yet) but for reparse if we don't do this it
means we will re-generate while reparsing which is just wrong

Fixed a few fragment bugs

Added the code to reformat the format nodes before reparsing so
that we can reparse a generated model... but ultimately because we
eliminated the generation from the idea in the test cases I was using
this code is not yet fully tested.  Need to add some unit tests but
trying to figure out how this feature will be used in the first place.