reparse and IDE fixes
- for reparse, added addReparseResultToParent to mirror the regular parse
routine

- changed approach to do completion for IDE - do not do the complete
partial reparse for command completion - instead just use the model
we are given and the original model (which has been started and so
has references).   So far, only do this for identifier expressions
but that's the biggest case.

- the auto-list wrapping logic in parselets is used for Try.statements
and Finally.statements.  For reparse, the list was not started by the
parent was and so we would not stop the children (and clear errors).

- when adding matching global names, only do that for layers which
are accessible to the refLayer