bug fixes and new feature for SymbolChoice
methodReference - for Boolean.class::cast we needed to first match classValueExpression, then type, then primary.

Need to test primary before type but I'm pretty sure there's a reason that's
not the right way.  If you look at the antlr grammar it's a complete patch
work.

Added caching of identifier since that might help with this method reference
pattern

Fixed bug with EnumConstant with interfaces

For SymbolChoice - supports excluded value - e.g. %> to avoid matching the
modulo operator %.   It's not usually an issue but for partial values we
are finding that 5 %> matches to 5 % in a binary expression.  It's better to
quickly disambiguate this case when you are in the template language, clearly
%> takes priority.  We could change this in the grammar with new nodes but
it makes it more awkward to change the grammar and will be less efficient
at runtime due to the extra parselets created.

reparse fix, I think the one for truncated files