type parameter fixes
tweak to setInferredType - need to call it except in the
deepCopy method from start

method reference disambiguation for how we choose when both
an instance method and a regular method match the reference method.
Ultimately we probably need a "pickMoreSpecificMethod" algorithm
that can handle the mismatching parameter lists (since one of the
methods moves the first param-type to the instance type).

fix for null return from resolveTypeParameter

fix for extractTypeParameter - use the right index and handle the
case where there's no type parameters in the final type