reparse fixes for schtml
also improvements to schtml incremental parsing

added acceptTree operator - so we consistently run the accept methods
on all parse nodes even for cached or reparsed results.

performance improvements to grammar for partial values.  do not parse
trees with partial values unless the start tag is fully complete

restoreCurrentIndex to preserve the semantic context - slightly different
than resetCurrentIndex which does not.

do not parse an html tree tag for an incomplete tag, mostly because of the
performance problems this creates.   Instead, it's parsed as a simple tag until it's
complete, then gets parsed as a tree tag.  Another way to handle this is to scan ahead
and see if we might possibly have a close tag?

we also require <tagName attributes to parse before we consider that a partial value for a tag.
this eliminates < as a partial value and a tag - instead just an error character